refactor(tangle-dapp): Move pie chart to `AllocationChart`
