+++ apps/tangle-dapp/containers/ManageProfileModalContainer/RolesInput.tsx
@@ -11,8 +11,8 @@ import usePolkadotApiRx from '../../hooks/usePolkadotApiRx';
 import { ServiceType } from '../../types';
 import { getChipColorByServiceType } from '../../utils';
 import { formatTokenBalance } from '../../utils/polkadot/tokens';
+import { getServiceChartColor } from './AllocationChart';
 import BaseInput from './BaseInput';
-import { getServiceChartColor } from './IndependentAllocationStep';
 
 export type RolesInputProps = {
   title: string;
